+++ packages/pieces/convertkit/src/lib/common/constants.ts
@@ -1,1 +1,2 @@
 export const CONVERTKIT_API_URL = 'https://api.convertkit.com/v3';
+export const WEBHOOK_BASE_OVERRIDE = 'https://example.ngrok.dev';
